+++ packages/implementations/react/src/atomic/Title/src/title.stories.tsx
@@ -15,6 +15,7 @@ type Story = StoryObj<typeof Title>;
 
 export const basic: Story = {
   args: {
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -22,6 +23,7 @@ export const basic: Story = {
 export const h1: Story = {
   args: {
     as: "h1",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -29,6 +31,7 @@ export const h1: Story = {
 export const h2: Story = {
   args: {
     as: "h2",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -36,6 +39,7 @@ export const h2: Story = {
 export const h3: Story = {
   args: {
     as: "h3",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -43,6 +47,7 @@ export const h3: Story = {
 export const h4: Story = {
   args: {
     as: "h4",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -50,6 +55,7 @@ export const h4: Story = {
 export const h5: Story = {
   args: {
     as: "h5",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };
@@ -57,6 +63,7 @@ export const h5: Story = {
 export const h6: Story = {
   args: {
     as: "h6",
+    color: "$primary",
     children: "Lorem ipsum dolor sit amet",
   },
 };